==> src/00_load_st_data.Rmd <==
---
title: "Load Spatial data"
date: "`r format(Sys.time(), '%d %m %Y')`"
format:
  html:
    embed-resources: true
    code-fold: show
params:
  fig.path: "./Figures/00/"
editor_options: 
  chunk_output_type: console
---

```{r background-job, eval=FALSE, include=FALSE}
source("../bin/render_with_jobs.R")

# quarto
# render_html_with_job(out_dir = lab_dir)
# fs::file_move(path = file, new_path = paste0(lab_dir, file))

# currently using quarto for github and kniter for html due to source code option 
render_git_with_job(fig_path = "./Figures/00/")
# Changes the figure path from ./Figures/ to ../Figures/:
system2(command = "sed", stdout = TRUE,
        args = c("-i", "''","-e", 's/src=\\"\\./src=\\"\\.\\./g',
                 paste0("./md_files/", basename("./00_load_st_data.md"))))

# kniter
knit_html_with_job(out_dir = "../lab_book/00_load_st_data/", fig_path = "./Figures/00/")
```

```{r setup, echo=FALSE, message=FALSE, warning=FALSE}
knitr::opts_chunk$set(
  fig.width   = 6.6929133858,
  fig.path    = params$fig.path,#"../Figures/",
  fig.align   = "center",
  message     = FALSE,
  warning     = FALSE,
  dev         = c("png"),
  dpi         = 300,
  fig.process = function(filename){
    new_filename <- stringr::str_remove(string = filename,
                                        pattern = "-1")
    fs::file_move(path = filename, new_path = new_filename)
    ifelse(fs::file_exists(new_filename), new_filename, filename)
  })

# setwd("/Users/vilkal/work/Brolidens_work/Projects/Spatial_DMPA/src/")
```

### Load libraries
```{r Load-libraries, message=FALSE, warning=FALSE}
##################
# LOAD LIBRARIES #
##################
library(tidyverse)
library(tidyseurat)
library(Seurat)
library(hdf5r)
# remotes::install_github("czarnewski/niceRplots",force=T)
library(niceRplots)

library(xml2) # loads the image 
library(sp)
library(terra)

source("../bin/help_functions.R")
source("../bin/spatial_visualization.R")

#################
# COLOR PALETTS #
#################
pal <- rep(c(RColorBrewer::brewer.pal(9,"Set1"),
         RColorBrewer::brewer.pal(9,"Pastel1"),
         RColorBrewer::brewer.pal(8,"Accent"),
         RColorBrewer::brewer.pal(8,"Set2"),
         RColorBrewer::brewer.pal(8,"Pastel2") ,
         scales::hue_pal()(8)),99)
# scales::show_col(pal)
```

### Load Visium data
```{r Load-data}
#########
# PATHS #
#########
m_path <- "https://www.ncbi.nlm.nih.gov/geo/download/?acc=GSE217237&format=file&file=GSE217237%5Fmetadata%2Ecsv%2Egz"
input_dir <- "../data/spatial_data"
result_dir <- "../results/00_load_st_data/"
if( isFALSE(dir.exists(result_dir)) ) { dir.create(result_dir,recursive = TRUE) }

#############
# LODA DATA #
#############
download.file(m_path, destfile = "../data/metadata.csv.gz")
meta <- read_csv(gzfile("../data/metadata.csv.gz"))
# meta <- read_csv("../data/Clinical_data/Clinical_data_Spatial_DMPA.csv")

h5_files <- list.dirs(path = input_dir,
                      full.names = T, recursive = T) %>%
            grep("P\\d\\d\\d$", ., value = TRUE) %>%
            set_names(., str_extract(., "P\\d\\d\\d$"))

# h5_files <- h5_files[names(h5_files) %in% sample_id]
# the new version of seurat v5 does not allow to read in any other than low res img
# to circumvent this rename the higres to low res and move the lowres file
image <- map(h5_files, 
             ~Read10X_Image(
               filter.matrix = T,
               image.dir = paste0(.x, "/spatial"),
               image.name = "tissue_lowres_image.png"))

# Read in h5 files and create Seurat Object
seuratObj_list <- pmap(list(h5_files, image, names(h5_files)),
                       ~Load10X_Spatial(
                         filename = "filtered_feature_bc_matrix.h5",
                         filter.matrix = T,
                         assay = "RNA",
                         data.dir = ..1,
                         image =  ..2,
                         slice = ..3)) 

sample_id <- c("P118", "P080", "P031", "P105", "P097", "P108", "P114", "P107") %>% set_names()
```

### Tidy up the seurat object
```{r, Clean-SeuratObj}
####################################
# RENAME SAMPLES, SPOTS AND IMAGES #
####################################
seuratObj_list <- seuratObj_list %>%
  imap(., ~AddMetaData(object = .x, 
                       metadata = rep(.y, length(Idents(.x))), 
                       col.name = "orig.ident")) %>%
  map(.,  ~SetIdent(., value = .@meta.data$orig.ident)) %>%
  imap(., ~RenameCells(.x, 
                      new.names = paste0(.y,"_", gsub("-.*","",colnames(.x[["RNA"]])))) ) %>%
  imap(., ~ {.x@images <- set_names(.@images,.y); .x})

##################
# MERGE SAMPLES #
#################
# Merge datasets into one single seurat object

DATA  <- merge(seuratObj_list[[1]], y = seuratObj_list[2:length(seuratObj_list)])
# DATA <- JoinLayers(DATA) # Seurat version 5.0
```

### Load morphology annotation
```{r Manual-annot}
##########################
# ADD MANUAL ANNOTATION #
#########################
# scale.factors <- map(DATA@images, ~pluck(.x, "scale.factors")$hires)
# img_coord <- map(DATA@images, ~pluck(.x, "coordinates"))
#   
# img_coord <- img_coord %>%
#   map2(., scale.factors, 
#        ~mutate(.x, imagecol = .$imagecol * .y,
#                    imagerow = .$imagerow * .y) )
  
a2 <- map(sample_id, ~xml2::as_list( read_xml( paste0( input_dir,"/",.x,"/",.x,".svg")) ) )
dd <- map(a2, ~as.numeric( strsplit( attributes(.x$svg)$viewBox , " ")[[1]] ))
dd2 <- map(sample_id, ~dim(DATA@images[[.x]]@image))

# sample_id <- "P097"
# a2 <- a2[["P097"]]
# dd <- dd[["P097"]]
# dd2 <- dd2[["P097"]]

# add image coordinates to the seurat object
get_img_coord <- function(DATA, sample_id){
  img_coord <- DATA@images[[sample_id]]@coordinates
  DATA@images[[sample_id]]@coordinates <<- img_coord 
}

# add manual spatial annotation
get_sp_annot <- function(a2, dd, dd2, sample_id){
  scale.factor <- DATA@images[[sample_id]]@scale.factors$hires
  img_coord <- DATA@images[[sample_id]]@coordinates
  
  id <- a2$svg %>%
    map_chr(., ~attr(.x,"id")) %>% 
    set_names(seq_along(.), .)
  
  annot_coord <- id %>%
    map(., ~get_shape(a2$svg[[.x]]) ) %>%
    map(., ~as_tibble(.x, .name_repair="unique"))  %>%
    # convert the view box from the svg to the pixel dimensions of the raster img:
    map(., ~mutate(.x, x = .x[[1]]*dd2[2]/dd[3],
                       y = .x[[2]]*dd2[1]/dd[4] )) %>%
    map(., ~rowid_to_column(., var = "path_idx")) %>%
    {. ->>  temp} %>%
    bind_rows(., .id="name") %>%
    group_by(., name) %>%
    mutate(elem_idx = cur_group_id()) %>% # group_indices(., name)
    ungroup() %>%
    mutate(colour = ifelse(grepl("fov|zoom|full_image",.$name),
                           "transparent", "black")) %>%
    list(.) %>%
    set_names(., sample_id[1])
  
  DATA@tools <<- append(DATA@tools, annot_coord )
  
  img_coord <- temp %>%
    list_modify("fov" = NULL, "full_image" = NULL, "zoom" = NULL) %>%
    compact() %>%
    imap(., ~mutate(img_coord, !!.y := sp::point.in.polygon(
         point.x = img_coord$imagecol*scale.factor,
         point.y = img_coord$imagerow*scale.factor,
         pol.x = .x$x,
         pol.y = .x$y )) ) %>%
    map(., ~select(.x, last_col())) %>%
    cbind(img_coord, .)
  
  DATA@images[[sample_id]]@coordinates <<- img_coord 
  
  sp_annot <- img_coord %>%
    rownames_to_column(., var = "barcodes") %>%
    mutate(across(7:ncol(.), ~ifelse(. == 0, NA, .)) ) %>%
    pivot_longer(., cols = 7:ncol(.), names_to ="sp_annot", values_to = "count") %>%
    filter(!(is.na(count))) %>%
    group_by(barcodes) %>%
    mutate(dupp = row_number()) %>%
    ungroup() %>%
    filter(., .$dupp == 1) %>%
    select(., barcodes, sp_annot)
  
  #DATA <<- left_join( DATA, sp_annot, by=c(".cell"="barcodes", "sp_annot"="sp_annot")) 
  
  return(list(coord=annot_coord, annot=sp_annot))
}
annot <- list(a2, dd, dd2, names(a2)) %>%
  pmap(., ~get_sp_annot(..1, ..2, ..3, ..4)) %>%
  set_names(., names(a2))

sp <- map(annot, 2) %>% bind_rows()
DATA <- left_join( DATA, sp, by=c(".cell"="barcodes")) 

DATA
```

```{r crop-img}
# adds a matrix with zeros outside the polygon boundary to DATA@misc$alpha
# when plotting it is added as a 4th dim to the image matrix  
# to specify alpha for each pixel in the image
mask_background.fun <- function(df, im, sampleid){
  # get polygon coordinates for individual polygons:
  df <- df %>%
    select( any_of(c("name","path_idx", "x", "y"))) %>% 
    filter(grepl("epi|SubMuc", .$name)) %>% 
    split(~name) 
  
  # Function to create SpatialPolygons from a dataframe
  createSpatialPolygons <- function(df) {
    # Convert dataframe to SpatialPointsDataFrame
    #coordinates(df) <- c("y","x")
    coordinates(df) <- c("x", "y")
    
    # Order the points based on 'path_idx'
    df <- df[order(df$path_idx, decreasing = F), ]
    #df <- df %>% arrange(desc(path_idx))
    
    # Create a SpatialPolygons object
    poly <- SpatialPolygons(list(Polygons(list(Polygon(df)), ID = df$name[1])))
    
    return(poly)
  }
  
  # Apply the function to each element of the list
  polygons_list <- map(df, ~createSpatialPolygons(.x))
  
  # Combine the SpatialPolygons into a single object
  combined_polygon <- do.call("rbind", polygons_list)
  
  # Convert it to a polygon (SpatVector)
  polygon <- vect(combined_polygon)
  
  # get image matrix and convert to rgb before rasterizing
  img <- matrix(
      rgb(im[,,1],im[,,2],im[,,3], im[4,,]* 1), nrow=dim(im)[1])
  raster <- rast(img)
  
  rasterized <- rasterize(polygon, raster)
  rasterized <- flip(rasterized, direction="vertical")
  
  # Plot the polygon
  # plot(polygon)

  # Crop the raster using the polygon
  cropped_raster <- mask(raster, rasterized, updatevalue=0)
  
  # Plot the cropped raster to verify
  # plot(cropped_raster)

  cropped_raster <- ifel(cropped_raster > 0, 1, cropped_raster)
  
  image_array <- as.array(cropped_raster)[,,1] # keep first dim only 
  #NA_idx <- which(is.na(image_array[,,1]))
  
  # minpulate image matrix
  #im <- replace(im, list = which(image_array != 0)), values = 1)
  DATA@misc$alpha <<- append(DATA@misc$alpha, set_names(list(image_array), sampleid) )
}

walk(sample_id, ~mask_background.fun(DATA@tools[[.x]], DATA@images[[.x]]@image, .x) )
```

```{r crop-by-boundary, eval=FALSE}
# can identify the boundary of the tissue and use it to crop
# it gives a more pixelated border than when using a hand drawn polygon as in the function
# above "mask_background.fun()"
find_boundary.fun <- function(im){
  img <- matrix(
      rgb(im[,,1],im[,,2],im[,,3], im[4,,]* 1), nrow=dim(im)[1])
  img <- rast(img)
  #plot(img)
  
  i <- im[,,1]+im[,,2]+im[,,3]
  #i <- replace(im[,,1], list = which(im[,,1] > .7), values = 0)

  raster <- rast(i)
  # plot(raster)
  raster <- ifel(raster > 2.9, NA, raster)
  r <- sieve(raster, directions=8, threshold=500)
  # plot(r)
  
  rc2 <- classify(r, matrix(c(0, NA,
                              2, 1,
                              3, 1), ncol=2, byrow=TRUE), 
                  include.lowest=TRUE, brackets=F)
  plot(rc2)
  
  # Extract only the boundary
  bo <- boundaries(rc2, inner=F, classes=T)
  # plot(bo)
  
  # Convert SpatRaster to SpatialPolygonsDataFrame
  boundary_polygon <- as.polygons(bo) 
  # plot(boundary_polygon)
  
  # Crop the raster using the polygon
  cropped_raster <- mask(img, boundary_polygon, updatevalue=NA)
  plot(cropped_raster)
  
  return(boundary_polygon)
}

boundary_polygon <- find_boundary.fun(DATA@images[["P118"]]@image)
```

### Add meta data
```{r add-groups}
meta <- meta %>%
  select(orig.ident="ID", groups=Contraception) %>%
  mutate(groups = ifelse(.$groups=="no HC", "ctrl", .$groups))

DATA <-  DATA %>%
  mutate(sp_annot2 = .$sp_annot) %>%
  mutate(sp_annot = ifelse(grepl("epi", .$sp_annot2), "epi", 
                           ifelse(grepl("SubMuc", .$sp_annot2), "SubMuc", .$sp_annot2 ))) %>%
  left_join(., meta, by="orig.ident") %>%
  select(groups, sp_annot, everything())

```

```{r 00a_plot_tissue, fig.height=8, fig.width=5}
# dev.new(height=8, width=5, noRStudioGD = TRUE)
# dev.new(width=7, height=3.5, noRStudioGD = TRUE)
plot_spatial.fun(DATA, 
      sampleid=sample_id,
      sp_annot = F, 
      alpha = 0,
      geneid = "nFeature_RNA",
      zoom = "zoom",
      ncol = 2,
      annot_line = 1,
      img_alpha = 1,
      point_size = 0) +  theme(legend.position = "none") # theme_nothing() #

```

```{r 00b_plot_sp_annot, fig.height=7, fig.width=6}
# dev.new(height=7, width=6, noRStudioGD = TRUE)
# dev.new(height=3.5, width=8, noRStudioGD = TRUE)
plot_st_meta.fun(DATA,  
          assay="RNA",
          feat = "sp_annot",
          zoom = "zoom",
          ncol = 2,
          annot_line = .1,
          img_alpha = 0,
          point_size = 0.9
        )
```

### Identify spots with missing morphology annotation
```{r Manuall-annotation }
##########################
# SP MANUAL ANNOTATION #
#########################
keep_epi <- c("P097_CTATTTGCTTGGAGGA", "P031_ATAGAGTTATCAACTT", "P031_GCGGAGAGGGAGAACG", "P031_CCTATCTATATCGGAA", "P031_GGATCTTGACTCAACC","P031_GCCCTAGCCGTCGCGA", "P031_AGCTCTTTACTCAGTT", "P031_AGTCAACACCACCATC", "P031_CGAACCCGCATGCGTC", "P080_TCGAGCCAGGCAGGCC", "P080_AACCCGACAACCCGTG","P080_GGCCGTTTGGGTTTCA","P080_CATTTGAGTGGTACGT", "P080_TTGGTCACACTCGTAA", "P080_CGAACCCGCATGCGTC", "P080_TTGCTGATCATGTTCG", "P080_CGAGACCCTAGAGTGT", "P080_CCAGCCTGGACCAATA", "P080_CGCATGGTGCGATGCT", "P080_ATAGACAACGGGACCT", "P080_CTCATTAACGTTGCCC", "P097_CGAGTTCTGTCCCACC", "P097_GTATGAAATTTCACTC", "P097_AGCAACCGAAAGTAAT", "P097_GAAGCCACTGATTATG", "P097_GCACTGCCTACCTTTA", "P097_CACAGCACCCACGGCA", "P097_AGTTTGGCCAGACCTA", "P097_ACAGAACTGAGAACAA", "P097_GCTTTCAGAGGAGGTG", "P107_ACATCCCGGCCATACG", "P107_ACGCAAACTAATAGAT","P107_ACTTGACTCCCTCTTT", "P107_GATCTTGGAGGGCATA", "P105_AAGACTGCAAGCTACT", "P105_ACACGGGAACTTAGGG", "P105_CACATTCTTTCGATGG", "P105_CCCGACCATAGTCCGC", "P105_CCTATGGGTTACCGTC", "P105_CCTCTAATCTGCCAAG", "P105_CGAGGCTAAATATGGC", "P105_CGTTTCACTTCGGGCG", "P105_GCGCTAATTGAATAGA", "P105_GATATGCGGTAGCCAA", "P105_GGCTCTGCTCCAACGC", "P105_GGGCTGCCTAGGGCGA", "P105_GGCTCTGCTCCAACGC", "P105_GGGCTGCCTAGGGCGA", "P105_GGTTTACAATCTCAAT", "P105_TGTGGCGGGCTTCTGG", "P105_TTAATCAGTACGTCAG", "P105_TTCATGGCGCAACAGG", "P105_TTGGGACACTGCCCGC", "P114_ACCTGCGTGTCATGTT", "P114_CAAGGATCGCATGTTC", "P114_AATGTTGTCGTGAGAC", "P108_ATACGCCGGCGAAACC", "P108_CACGTCGGCAACCTCT", "P108_TCGCCGAAGTTGCGTC", "P108_TTGATTAGCTGTTTCT", "P118_AACCCGACAACCCGTG", "P118_TGAGCCATACAGTCTC", "P118_ATGGGACCTGCTGAAC", "P118_AGCTAACAAGCAATGT", "P118_TTGATTAGCTGTTTCT", "P118_CGTGTCTCGTTACGAC", "P118_GATCAACATAAAGGGA", "P118_TAACAGCGTTTGTGCT", "P118_TGAGCCATACAGTCTC", "P118_GATTACTGAATTTGGG", "P118_TCCAACTTTAAATTCT", "P118_CGATCCTCGCAACATA", "P118_AACCCGACAACCCGTG", "P118_TCTCTTACCGCGAACC", "P118_TATGTAAAGTGCTTAA", "P118_GAAGTTTCCACTCAAT", "P118_TAGTCCCGGAGACCAC", "P118_CGGCCAGAGCGACCAT", "P118_ATAAAGGCTCGGTCGT", "P118_TATTCGTGCCAGAATA", "P118_GGGAGTTAATGAGGCG", "P118_CCGGGCGGTCTCGTCA",
"P118_CTAAATCCGGTGTACA") 
keep_SubMuc <- c("P105_TGACATCGAGCGGACC", "P118_GTATCAAACGTTAGCT", "P097_AGGTGGTGACCTTCGC", "P097_AGCCCGGCATTAGAGG", "P031_CTAGTTGGGCCCGGTA","P031_TCTACCGTCCACAAGC", "P031_AAATGGCCCGTGCCCT", "P031_AATCTGGCTTTCTAGT", "P031_AAATGGCCCGTGCCCT", "P031_AATCTGGCTTTCTAGT", "P080_AGGCATTGTCGTAGGG", "P097_GGCAGCAAACCTATGC", "P097_ACAAAGCATGACCTAG", "P097_CCTCCTGTTGTGTCGT", "P097_CGTCGGATAGTGTTGA", "P107_TCTTGATGCGTAGCGA", "P105_ACGATCATCTTGTAAA", "P105_AGATATAATACGACTA", "P105_CTAGGTCTGAAGGAAT", "P114_AAGGGTTTGATTTCAG", "P114_GCACGTGGTTTACTTA", "P114_TAGGCCTATATAGTCT", "P107_TCTTCGATACCAATAA", "P108_GTGCGAAATCGAACAC", "P118_TCGAGACCAACACCGT", "P118_TCGGAGTACATGAGTA", "P118_TATTCAATTCTAATCC", "P118_AGGAGGCCTTCGCGCG",
"P118_ACAATTGTGTCTCTTT", "P118_ACTGAAACGCCGTTAG", "P118_CCCGGGTCGTTCAGGG", "P118_CCCTCCTCGCTCGTAT", "P118_GCGCCTCCCACTCCGA", "P118_GCTAACTGAAGTCTGA",
"P118_TGTGGCGGGCTTCTGG") 

###############################
# IDENTIFY MISSING ANNOTATION #
###############################
df <- map(sample_id, ~pluck(DATA@images, .x, "coordinates")) %>%
  bind_rows() %>%
  #cbind(.,as_tibble(select(DATA, filt))) %>%
  cbind(.,as_tibble(select(DATA, orig.ident))) %>%
  cbind(.,as_tibble(select(DATA, nCount_RNA))) %>%
  cbind(.,as_tibble(select(DATA, nFeature_RNA))) %>%
  cbind(.,as_tibble(select(DATA, sp_annot))) %>%
  rownames_to_column(var = "barcode") %>%
  as_tibble() %>%
  select(-starts_with(c("epi_", "SubMuc_")))

f <- df %>% 
  filter(orig.ident == "P097" & is.na(.$sp_annot)) #%>%
  #filter(orig.ident == "P118" & filt == "keep")

# dev.new(width=10, height=10, noRStudioGD = TRUE)
# dev.new(width=5, height=3, noRStudioGD = TRUE)
# DATA %>%
#   mutate(sp_annot = case_when(colnames(DATA) %in% keep_epi ~ 'epi',
#                               colnames(DATA) %in% keep_SubMuc ~ 'SubMuc',
#                               TRUE ~ .$sp_annot)) %>%
#   mutate(filt = case_when(is.na(.$sp_annot) & nCount_RNA < 500 ~ 'filt',
#                           colnames(DATA) %in% filt ~ 'filt',
#                           TRUE ~ 'keep')) %>%
#   filter(grepl("P097_TAGAATAGCCGATGAA", `.cell`)) %>%
#   plot_st_meta.fun(.,  
#           assay="RNA",
#           feat = "filt",
#           zoom = "full_image",
#           ncol = 2,
#           annot_line = .1,
#           img_alpha = 0,
#           point_size = 1
#         )
```

### Add missing morphology annotation
```{r 00c_plot_new_sp_annot, fig.height=8, fig.width=6}
# dev.new(width=10, height=10, noRStudioGD = TRUE)
DATA <-  DATA %>%
  mutate(sp_annot = case_when(colnames(DATA) %in% keep_epi ~ 'epi',
                              colnames(DATA) %in% keep_SubMuc ~ 'SubMuc',
                              TRUE ~ .$sp_annot))

DATA %>%
  plot_st_meta.fun(.,  
          assay="RNA",
          feat = "sp_annot",
          zoom = "zoom",
          ncol = 2,
          annot_line = .1,
          img_alpha = 0,
          point_size = 0.8
        )
```

## Plot spots to be removed
```{r 00d_plot_spots_to_remove, fig.height=8, fig.width=6}
# dev.new(width=10, height=10, noRStudioGD = TRUE)

DATA <-  DATA %>%
  mutate(filt = case_when(is.na(.$sp_annot)  ~ 'filt',
                          TRUE ~ 'keep')) 
DATA %>%
  plot_st_meta.fun(.,  
          assay="RNA",
          feat = "filt",
          zoom = "full_image",
          ncol = 2,
          annot_line = .1,
          img_alpha = 0,
          point_size = .8
        )
```

### Remove NA spots
```{r remove-NA-spots}
dim(DATA)
# Filter spots outside manual annotation
DATA <- DATA[, !(is.na(DATA$sp_annot))]
DATA$filt <- NULL

dim(DATA)
DATA
```

```{r 00e_final-sp_annot, fig.height=7, fig.width=6}
# dev.new(height=7, width=6, noRStudioGD = TRUE)
# dev.new(height=3.5, width=8, noRStudioGD = TRUE)
DATA %>%
  #filter(orig.ident == "P118" | orig.ident == "P097") %>%
  plot_st_meta.fun(.,  
          assay="RNA",
          feat = "sp_annot",
          zoom = "zoom",
          ncol = 2,
          annot_line = .1,
          annot_col = "black",
          img_alpha = 0,
          point_size = 0.8
        )
```

## Save seurat object
```{r save-SeuratObject}
##################################
# SAVE INTERMEDIATE SEURAT OJECT #
##################################
saveRDS(DATA, paste0(result_dir,"seuratObj_merged.RDS"))
# DATA <- readRDS(paste0(result_dir,"seuratObj_merged.RDS")) # seuratObj_merged_seurat_v5.RDS
```

### Session info
```{r}
sessionInfo()
```

```{r Paulos-read-data, eval=FALSE, include=FALSE}
Read10X_Image2 <- function (image.dir, filter.matrix = TRUE, ...) { 
  image <- png::readPNG(source = file.path(image.dir, "tissue_hires_image.png"))
    scale.factors <- jsonlite::fromJSON(txt = file.path(image.dir, "scalefactors_json.json"))
    tissue.positions <- read.csv(file = file.path(image.dir, 
        "tissue_positions_list.csv"), col.names = c("barcodes", 
        "tissue", "row", "col", "imagerow", "imagecol"), header = FALSE, 
        as.is = TRUE, row.names = 1)
    if (filter.matrix) {
        tissue.positions <- tissue.positions[which(x = tissue.positions$tissue == 
            1), , drop = FALSE]
    }
    unnormalized.radius <- scale.factors$fiducial_diameter_fullres * 
        scale.factors$tissue_lowres_scalef
    spot.radius <- unnormalized.radius/max(dim(x = image))
    return(new(Class = "VisiumV1", image = image, scale.factors = scalefactors(spot = scale.factors$tissue_hires_scalef, 
        fiducial = scale.factors$fiducial_diameter_fullres, hires = scale.factors$tissue_hires_scalef, 
        scale.factors$tissue_lowres_scalef), coordinates = tissue.positions, 
        spot.radius = spot.radius)) }

# DEFINE VARIABLES AND PATHS TO THE DATASET #
PATH <- "../data" #"../data/spatial_data/"
filelist <- list.dirs(paste0(PATH,"/spatial_data"),recursive = F)
# LOAD COUNTS AND CREATE SEURAT OBJECT #
DATA_list <- lapply(filelist,function(i){
  x <- sub(".*[/]","",i)
  DATA <- Read10X_h5(paste0(i,"/raw_feature_bc_matrix.h5"))
  colnames(DATA) <- paste0(sub("[-].*","",colnames(DATA)),"_",x)
  DATA <- CreateSeuratObject(counts = DATA,project = x)
})


# 
DATA <- merge(DATA_list[[1]],DATA_list[-1])
dim(DATA)


# LOAD IMAGES INTO SEURAT OBJECT #
for(i in filelist){
  x <- sub(".*[/]","",i)
  image <- Read10X_Image2(image.dir = paste0(i,"/spatial"), filter.matrix = FALSE)
  rownames(image@coordinates) <- paste0(sub("[-].*","",rownames(image@coordinates)),"_",x)
  image@coordinates <- image@coordinates[ colnames(DATA)[DATA$orig.ident == x] , ]
  DefaultAssay(object = image) <- "RNA"
  DATA[[x]] <- image
}

  
rm(DATA_list)
rm(image)
```

```{r Paulos-manual-annot, eval=FALSE, include=FALSE}
##########################
# ADD MANUAL ANNOTATION #
#########################
DATA$sp_annot <- "other"
#DATA$orig.ident <- factor(DATA$orig.ident)
# DATA <- DATA[,DATA$orig.ident == "P107"]
# red <- "P107"

for(red in names(DATA@images) ){
  red <-  "P080"#
  coo <- DATA@images[[red]]@coordinates
  coo$imagecol <- coo$imagecol*DATA@images[[red]]@scale.factors$hires
  coo$imagerow <- dim(DATA@images[[red]]@image)[1] - coo$imagerow*DATA@images[[red]]@scale.factors$hires
  
  sum(rownames(coo) %in% rownames(DATA@meta.data))
  
   plot_spatial_feat(
    DATA, red = red,
    feat =  "ACTB",
    main="",assay="RNA",pch=16,cex=.5,
    plot_tissue = F)
  
  path <-  paste0( input_dir,"/",red,"/",red,".svg")

  svg <- readChar(path, file.info(path)$size)
  x <- xml2::read_xml(svg)
  a2 <- as_list(xml_child(x, 1))
  
  # path to svg file defining the regions:
  a2 <- xml2::as_list( read_xml( paste0( input_dir,"/",red,"/",red,".svg")) ) 
  dd <- as.numeric( strsplit( attributes(a2$svg)$viewBox , " ")[[1]] )
  dd2 <- dim(DATA@images[[red]]@image)
  
  for(i in order( sapply(a2$svg,function(x) attr(x,"id")) ) ){
    #i <- "epi_1"
    #i <- "SubMuc"
    #i <- 5
    if(!(attr(a2$svg[[i]],"id") %in% c("full_image","fov")) ){
      df <- get_shape(a2$svg[[i]])
      id <- attr(a2$svg[[i]],"id")
      
      polygon((df[,1])*dd2[2]/dd[3], (dd[4]-df[,2])*dd2[1]/dd[4] ,xpd=T)
     
      res <- (sp::point.in.polygon(
        point.x = coo$imagecol,
        point.y = coo$imagerow,
        pol.x = (df[,1])*dd2[2]/dd[3],
        pol.y = (dd[4]-df[,2])*dd2[1]/dd[4] ))>0
      
      DATA@images[[red]]@coordinates <- DATA@images[[red]]@coordinates %>%
        mutate( !!id := ifelse(res==TRUE, 1, 0))
      
      DATA@meta.data[rownames(coo)[res],"sp_annot"] <- attr(a2$svg[[i]],"id")
      print(path)
    }
  }
}
DATA$sp_annot <- factor(DATA$sp_annot)

plot_spatial_meta(
    DATA, red = red,
    feat =  "sp_annot",
    main="",assay="RNA",pch=16,cex=.5,
    plot_tissue = T)
```
