==> src/02_integrate_st_data.Rmd <==
---
title: "Integrate spatial data"
date: "`r format(Sys.time(), '%d %B, %Y')`"
format:
  html:
    embed-resources: true
    code-fold: show
params:
  fig.path: "`r paste0(params$fig.path)`" #./Figures/
editor_options: 
  chunk_output_type: console
---

```{r background-job, eval=FALSE, include=FALSE}
source("../bin/render_with_jobs.R")

# quarto
# render_html_with_job(out_dir = lab_dir)
# fs::file_move(path = file, new_path = paste0(lab_dir, file))

# currently using quarto for github and kniter for html du to source code option 
render_git_with_job(fig_path = "./Figures/02/")
# Change the figure path from ./Figures/03/ to ../Figures/03/:
system2(command = "sed", stdout = TRUE,
        args = c("-i", "''","-e", 's/src=\\"\\./src=\\"\\.\\./g',
                 paste0("./md_files/", basename("./02_integrate_st_data.md"))))

# kniter
knit_html_with_job(out_dir = "../lab_book/02_QC_integrate_st_data/", fig_path = "./Figures/02/")
```

```{r setup, include=FALSE}
knitr::opts_chunk$set(
  fig.width = 6.6929133858,
  fig.path    = params$fig.path,
  dev         = c("png"),
  dpi         = 300,
  fig.align   = "center",
  message     = FALSE,
  warning     = FALSE,
  fig.process = function(filename){
      new_filename <- stringr::str_remove(string = filename, 
                                        pattern = "-1")
      fs::file_move(path = filename, new_path = new_filename)
      ifelse(fs::file_exists(new_filename), new_filename, filename)})
# setwd("/Users/vilkal/work/Brolidens_work/Projects/Spatial_DMPA/src")
```

### Load data and libraries
```{r Load-Library-and-data}
##################
# LOAD LIBRARIES #
##################
library(tidyverse)
library(Seurat)
library(tidyseurat)
library(cowplot)
library(harmony)
source("../bin/plotting_functions.R")

#########
# PATHS #
#########
input_dir <- "../results/01_QC_st_data/"
result_dir <- "../results/02_integrate_st_data/"
if( isFALSE(dir.exists(result_dir)) ) { dir.create(result_dir,recursive = TRUE) }

#############
# LODA DATA #
#############
DATA <- readRDS(paste0(input_dir,"seuratObj_filtered.RDS"))

```

### Identify Highly Variable Genes (HVG) across samples
```{r Find-HVG}
################################
# SPLIT INTO SEPERATE DATASETS #
################################
DATA_nested <- DATA %>%
  mutate(batch = orig.ident) %>%
  nest(data = -batch) %>%
  mutate(data = imap(
    data, ~ .x %>%
      NormalizeData(., normalization.method = "LogNormalize", 
                    verbose = FALSE) %>%
      FindVariableFeatures(selection.method = "vst", 
                           nfeatures = 2000, 
                           verbose = FALSE) )) %>%
  mutate(data = setNames(.[["data"]], .$batch))

#########################################
# FIND HIGLY VARIABLE GENES PER DATASET #
#########################################
hvgs_heat <- DATA_nested %>%
  .$data %>%
  map(., ~ .x@assays$RNA@var.features) %>%
  ( function(x){unique(unlist(x)) ->> hvgs_all; return(x)} ) %>%
  # intersect across all samples:
  ( function(x){Reduce(intersect, x) ->> hvgs; return(x)} ) %>% 
  imap_dfc(., ~hvgs_all %in% .x, .id=.y) %>%
  mutate(rownames = hvgs_all) %>%
  column_to_rownames(var = "rownames")

# choose the hvg present in at least two samples:
hig_var <- rownames(hvgs_heat)[rowSums(hvgs_heat)>2]

# remove all VDJ-genes from list of HVG
remove <- str_subset(hig_var, "^IGH|^IGK|^IGL|^TRA|^TRB|^TRD|^TRG")
hig_var <- setdiff(hig_var, remove)
```

### Heatmap of HVG in all samples
```{r 02a_HVG_heatmap, fig.width=10}
pheatmap::pheatmap(t(hvgs_heat * 1), cluster_rows = F, color = c("grey90", "grey20"))
```


### Integration
```{r Integration}
############
# HARMONY #
###########
DATA <- DATA %>%
  NormalizeData(verbose = FALSE) %>%
  FindVariableFeatures(selection.method = "vst",
                      nfeatures = 4000,
                      verbose = FALSE) %>%
  ScaleData(verbose = FALSE, features = hig_var ) %>%
  RunPCA(verbose = FALSE, npcs = 50) %>%
  RunUMAP(dims = 1:50,
          n.components = 2L,
          n.neighbors = 10,
          min.dist = .1,
          spread = .3) 

DATA <- DATA %>%
  RunHarmony(group.by.vars = "orig.ident", 
             reduction = "pca",
             dims.use = 1:50, 
             assay.use = "RNA") #%>%

DATA <-   DATA %>%
  RunUMAP(dims = 1:50, 
          n.neighbors = 10,
          min.dist = .1,
          spread = 1,
          repulsion.strength = 1,
          negative.sample.rate = 10,
          n.epochs = 100,
          reduction = "harmony",
          reduction.name = "umap_harmony")
```

### Alternative graph based UMAP
```{r alternative-way-of-UMAP}
integrated <- DATA@reductions$harmony@cell.embeddings
ann <- RcppHNSW::hnsw_build(as.matrix(integrated), distance = "cosine")
knn <- RcppHNSW::hnsw_search(as.matrix(integrated) , ann = ann , k = 15)

UU2 <- uwot::umap(X = NULL,
                 nn_method =  knn,
                 n_components = 2,
                 ret_extra = c("model","fgraph"),
                 verbose = T,
                 min_dist = 0.1,
                 spread = .3,
                 repulsion_strength = 1,
                 negative_sample_rate = 10,
                 n_epochs = 150,
                 n_threads = 8)
dimnames(UU2$embedding) <- list(colnames(DATA),paste0("umap_harmony_knn_", 1:2))
DATA@reductions[["umap_harmony_knn"]] <- CreateDimReducObject(embeddings = UU2$embedding, 
                                                              key = "umap_harmony_knn_")
colnames(DATA@reductions$umap_harmony_knn@cell.embeddings) <- paste0("umap_harmony_knn_", 1:2)
```

```{r, 02b_UMAP_options, fig.asp=5/10}
res <- c("umap_harmony", "umap_harmony_knn")
p <- map(res, ~plot_clusters.fun(DATA, red=.x, cluster="orig.ident", lable=FALSE, txt_size = 7))
plot_grid(ncol = 2, 
          plotlist = p)
```

### Plot before and after integration
```{r 02c_Plot_dim_reduction, fig.height=6}
#  dev.new(height=6, width=6.6929133858, noRStudioGD = TRUE)
res <- c("PC", "harmony", "UMAP", "umap_harmony")
title <- c("PCA raw data", "PCA Harmony integrated", "UMAP raw data", "UMAP Harmony integrated")
p <- map2(res, title, 
          ~plot_clusters.fun(DATA, 
                             cluster="orig.ident", txt_size = 9,
                             red=.x, lable=FALSE, title=.y))
plot_grid(ncol = 2, 
         plotlist = p)
```

### Plot marker genes
```{r 02d_plot_marker_genes, fig.height=3, fig.width=8}
#  dev.new(height=3, width=8, noRStudioGD = TRUE)
################################
# VISUALIZE EXPR. OF KEY GENES #
################################
# col <- c("grey90","grey80","grey60","navy","black")
col <- c("#FFF5F0", "#FEE0D2", "#FCBBA1", "#FC9272", "#FB6A4A", "#EF3B2C", "#CB181D", "#A50F15", "#67000D")
genes <- c("KRT1", "KRT15", "CDH1")
# genes <- c("CD8A", "SFRP2", "CD3E")
# genes <- c("CD8A", "MYOZ2", "CD3E", "EPCAM", "COL6A1", "CD4")

p <- map(genes, ~plot_genes.fun(DATA, .x, col = col, lable = FALSE))
plot_grid(ncol = 3, 
          plotlist = p)
```

## Save seurat object
```{r save-SeuratObj}
##################################
# SAVE INTERMEDIATE SEURAT OJECT #
##################################
saveRDS(DATA, paste0(result_dir,"seuratObj_integrated.RDS"))
# DATA <- readRDS(paste0(result_dir,"seuratObj_integrated.RDS"))
```

### Session info
```{r}
sessionInfo()
```

## Paulos base R code
```{r Paulos-HVG, eval=FALSE, include=FALSE}
DATA.list <- SplitObject(DATA, split.by = "sample_name")

for (i in 1:length(DATA.list)) {
    DATA.list[[i]] <- NormalizeData(DATA.list[[i]], verbose = FALSE)
    DATA.list[[i]] <- FindVariableFeatures(DATA.list[[i]], selection.method = "vst",
        nfeatures = 2000, verbose = FALSE)
}

hvgs_per_dataset <- lapply(DATA.list, function(x) {
    x@assays$RNA@var.features
})
# venn::venn(hvgs_per_dataset,opacity = .4,zcolor = scales::hue_pal()(3),cexsn
# = 1,cexil = 1,lwd=1,col='white',frame=F,borders = NA)

temp <- unique(unlist(hvgs_per_dataset))
overlap <- sapply(hvgs_per_dataset, function(x) {
    temp %in% x
})
rownames(overlap) <- temp
pheatmap::pheatmap(t(overlap * 1), cluster_rows = F, color = c("grey90", "grey20"))
```

```{r Paulos-harmony, eval=FALSE, include=FALSE}
hig_var <- rownames(overlap)[rowSums(overlap)>2]
DATA <- DATA %>%
  NormalizeData(verbose = FALSE) %>%
  FindVariableFeatures(selection.method = "vst",
                      nfeatures = 4000,
                      verbose = FALSE) %>%
  ScaleData(verbose = FALSE, features = hig_var ) %>%
  RunPCA(verbose = FALSE, npcs = 50) %>%
  RunUMAP(dims = 1:50, n.components = 2L) 

DATA <- RunHarmony(DATA, group.by.vars = "sample_name", reduction = "pca",
    dims.use = 1:50, assay.use = "RNA")

DATA <- RunUMAP(DATA, dims = 1:30, reduction = "harmony", reduction.name = "umap_harmony")
```

```{r Plot_dimreduction, eval=FALSE, fig.width=10, include=FALSE}
plot_grid(ncol = 2,
  DimPlot(DATA, reduction = "pca", group.by = "orig.ident")+NoAxes()+ggtitle("PCA raw_data"),
  DimPlot(DATA, reduction = "umap", group.by = "orig.ident")+NoAxes()+ggtitle("UMAP raw_data"),
  
  DimPlot(DATA, reduction = "harmony", group.by = "orig.ident")+NoAxes()+ggtitle("harmony integrated"),
  DimPlot(DATA, reduction = "umap_harmony", group.by = "orig.ident")+NoAxes()+ggtitle("UMAP harmony integrated")

```

```{r Paulos-plot, eval=FALSE, include=FALSE}
library(niceRplots)
genes <- c("CD8A", "MYOZ2", "CD3E", "EPCAM", "COL6A1", "CD4")
genes <- c("CDH1")
par(mfrow=c(2,2), mar=c(1,1,1,1))
DATA$cell_id[is.na(DATA$cell_id)] <- ""
#plot_meta(DATA, red = "umap_harmony", feat = "cell_id", cex = 0.5, label = T)
#plot_meta(DATA, red = "umap_harmony", feat = "p", cex = 0.5, label = T)
for(i in genes){
  plot_feat(DATA, red = "umap_harmony", feat = i, cex = 0.5)
}
grep("MYO", rownames(DATA), value = T)
```

```{r Paulos-code, eval=FALSE, include=FALSE}
##################################
# EVALUATE CLUSTERING RESOLUTION #
##################################
DATA <- FindNeighbors(DATA, reduction = "harmony", dims = 1:20, k.param = 20, prune.SNN = 1/15) 

for (res in c(0.1, 0.25, 0.5, 1, 1.5, 2)) {
    DATA <- FindClusters(DATA, graph.name = "RNA_snn", resolution = res, algorithm = 1)
}

library(niceRplots)
par(mfrow = c(1,2))
plot_meta(DATA, red = "umap_harmony", feat = "RNA_snn_res.0.25", cex = 0.5)
plot_meta(DATA, red = "umap_harmony", feat = "RNA_snn_res.1", cex = 0.5, label = T)

# plot_grid(ncol = 3, DimPlot(DATA, reduction = "umap", group.by = "RNA_snn_res.0.5") +
#     ggtitle("leiden_0.5"), DimPlot(DATA, reduction = "umap", group.by = "RNA_snn_res.1") +
#     ggtitle("leiden_1"), DimPlot(DATA, reduction = "umap", group.by = "RNA_snn_res.2") +
#     ggtitle("leiden_2"))

```
