==> src/05_DGE_clusters_st_data.Rmd <==
---
title: "Differential gene expression clusters"
date: "`r format(Sys.time(), '%d %m %Y')`"
format:
  html:
    embed-resources: true
    code-fold: show
params:
  fig.path: "`r paste0(params$fig.path)`" #./Figures/
  re.run: FALSE
editor_options: 
  chunk_output_type: console
---

```{r background-job, eval=FALSE, include=FALSE}
source("../bin/render_with_jobs.R")
# quarto
# render_html_with_job(out_dir = lab_dir)
# fs::file_move(path = file, new_path = paste0(lab_dir, file))

# currently using quarto for github and kniter for html due to source code option 
render_git_with_job(fig_path = "./Figures/05/")
system2(command = "sed", stdout = TRUE,
        args = c("-i", "''","-e", 's/src=\\"\\./src=\\"\\.\\./g',
                 paste0("./md_files/", basename("./05_DGE_clusters_st_data.md"))))

# kniter
knit_html_with_job(out_dir = "../lab_book/05_DGE_clusters_st_data", fig_path = "./Figures/05/")
```

```{r setup, include=FALSE}
knitr::opts_chunk$set(
  fig.width     = 6.6929133858,
  fig.path      = params$fig.path,#"../Figures/",
  fig.align     = "center",
  message       = FALSE,
  warning       = FALSE,
  dev           = c("png"),
  dpi           = 300,
  fig.process = function(filename){
    new_filename <- stringr::str_remove(string = filename,
                                        pattern = "-1")
    fs::file_move(path = filename, new_path = new_filename)
    ifelse(fs::file_exists(new_filename), new_filename, filename)
  }
  )
# setwd("/Users/vilkal/work/Brolidens_work/Projects/Spatial_DMPA/src")
```

## Load data and libraries
```{r Load-data}
##################
# LOAD LIBRARIES #
##################
library(tidyverse)
library(Seurat)
library(SeuratObject)
library(tidyseurat)
library(cowplot)
library(ggrepel)
library(niceRplots)
library(MAST)
library(scran)
library(openxlsx)

source("../bin/plotting_functions.R")
source("../bin/spatial_visualization.R")

#########
# PATHS #
#########
input_dir <- "../results/04_deconvolution_st_data/"
result_dir <- "../results/05_DGE_clusters_st_data/"
marker_dir <- "./marker_genes_clusters/"
if( isFALSE(dir.exists(result_dir)) ) { dir.create(result_dir,recursive = TRUE) }
if( isFALSE(dir.exists(marker_dir)) ) { dir.create(marker_dir,recursive = TRUE) }

#############
# LODA DATA #
#############
DATA <- readRDS(paste0(input_dir,"seuratObj_deconvolution_scdc.RDS"))
#DATA <- readRDS(paste0("../results/03_clustering_st_data/","seuratObj_clustered.RDS"))
epi_clus <- "^5$|^6$|^7|^9" # non-filt
```

```{r Functions}
########################
# SEURAT DGEs FUNCTION #
########################
# obj <- DATA_sub$data[[1]]
# clusters <- "Clusters"
DEGs_fun <- function(obj, clusters){
  obj <- SetIdent(obj, value = clusters)
  
  DATA_degs <- obj %>%
     FindAllMarkers(.,
                   test.use = "wilcox",
                   only.pos = F,
                   return.thresh = 1,
                   max.cells.per.ident = Inf,
                   logfc.threshold = -Inf,
                   assay = "RNA",
                   min.pct = -Inf)

  DEGs_table <- DATA_degs %>%
    as_tibble(.) %>%
    #map2(., comb$comb, ~mutate(.x, Combination = .y)) %>%
    #left_join(., Avg, by= c("gene")) %>%
    mutate(cluster = paste0(.$cluster, "_all")) %>%
    mutate(pct.diff = -.$pct.2-.$pct.1) %>%
    mutate(log.pct.diff = -log2(.$pct.2/.$pct.1))
  return(DEGs_table)
}

```

```{r Subset-data, include=TRUE, eval=FALSE}
####################################
# SUBSET SEURAT OBJECT PER CLUSTER #
####################################
# creating a subseted object with 25 spots per sampleID for each cluster
set.seed(1)
DATA_sub <- DATA %>%
  mutate(gr = .$groups) %>%
  mutate(ID = .$orig.ident) %>%
  mutate(sp = .$sp_annot) %>%
  nest(., data=-c(gr, orig.ident, sp)) %>%
  mutate(downsample = ifelse(.$sp == "epi", 25, 50)) %>%
  mutate(downsample_morf = ifelse(.$sp == "epi", 25, 25)) %>%
  mutate(subset = map2(data, .$downsample, ~filter(.x, .cell %in% WhichCells(., downsample = .y)))) %>%
  mutate(subset_morf = map2(data, .$downsample_morf, ~filter(.x, .cell %in% WhichCells(., downsample = .y)))) %>%
  mutate(across(c("data", "subset", "subset_morf"), ~map(., ~table(.x$Clusters)), .names = "n_{.col}")) %>%
  mutate(across(contains("n_"), ~set_names(.x, paste0(.data[["gr"]],"_",.data[["orig.ident"]])))) %>%
  mutate(across(contains("n_subset_morf"), ~set_names(.x, paste0(.data[["gr"]],"_",.data[["orig.ident"]],"_", .data[["sp"]])))) %>%
  select(-data) %>%
  {. ->> full_subset } %>% # [,-c("subset", "n_subset")]
  select(-contains("_morf")) %>%
  split(., ~sp) %>%
  c(., full_subset = list(full_subset))

saveRDS(DATA_sub, paste0(result_dir, "DATA_sub.RDS"))
```

```{r Load-DATA_sub}
DATA_sub <- readRDS(paste0(result_dir, "DATA_sub.RDS"))
```

```{r spot-distribution-in-subsetted-data}
##############
# EPITHELIUM #
##############
bind_cols(DATA_sub$epi$n_subset, "Clus" = paste0("**",names(table(DATA$Clusters)),"**")) %>%
  rowwise() %>% 
  mutate(DMPA_sum = sum(c_across(starts_with("DMPA_"))),
         ctrl_sum = sum(c_across(starts_with("ctrl_")))) %>%
  select(sort(colnames(.)[1:8]), everything()) %>%
  knitr::kable(., caption = "Distribution of epithelial spots per cluster per subject")

#############
# SUBMUCOSA #
#############
bind_cols(DATA_sub$SubMuc$n_subset, "Clus" = paste0("**",names(table(DATA$Clusters)),"**")) %>%
  rowwise() %>% 
  mutate(DMPA_sum = sum(c_across(starts_with("DMPA_"))),
         ctrl_sum = sum(c_across(starts_with("ctrl_")))) %>%
  select(sort(colnames(.)[1:8]), everything()) %>%
  knitr::kable(., caption = "Distribution of submucosal spots per cluster per subject")

###################
# SUBMUC VS EPI #
##################
bind_cols(DATA_sub$full_subset$n_subset_morf, "Clus" = paste0("**",names(table(DATA$Clusters)),"**")) %>%
  rowwise() %>% 
  mutate(SubMuc_sum = sum(c_across(ends_with("_SubMuc"))),
         epi_sum = sum(c_across(ends_with("_epi")))) %>%
  select(sort(colnames(.)[1:8]), everything()) %>%
  {. ->> tab} %>%
  knitr::kable(., caption = "Distribution of spots per cluster per subject")

colSums(tab[,c("SubMuc_sum", "epi_sum")]) %>%
  knitr::kable(., caption = "Total number of spots in epi and Submucosa")
```

```{r plot_clustering_on_tissue, fig.height=3.5, fig.width=7, include=FALSE, eval=FALSE}
plots <- DATA_sub %>%
  plot_st_meta.fun(.,  
            assay="RNA",
            feat = "Clusters",
            zoom = "zoom",
            ncol = 4,
            annot_line = .1,
            img_alpha = 0,
            point_size = 0.8
          )

```

### Run differential gene expression analysis
```{r DEGs, include=TRUE, eval=params$re.run}
#############################
# DGEs ONE AGAINST THE REST #
#############################
DEGs_table_epi <- DATA_sub$epi %>%
  unnest(subset) %>%
  filter(grepl(epi_clus, .$Clusters)) %>%
  {. ->> DATA_sub_e} %>%
  DEGs_fun(., "Clusters") 

DEGs_table_subMuc <- DATA_sub$SubMuc %>%
  unnest(subset) %>%
  filter(!(grepl(epi_clus, .$Clusters))) %>%
  {. ->> DATA_sub_s} %>%
  DEGs_fun(., "Clusters") 

DEGs_table <- bind_rows(DEGs_table_epi, DEGs_table_subMuc)

###################
# ADD ANNOTATION #
##################
ord1 <- c("Superficial","Upper IM","Lower IM","Basal","8","3","4","0","2","1","10")
ord2 <- c("6_all","9_all","7_all","5_all","8_all","4_all","3_all","2_all","1_all","0_all","10_all")
ord2 <- c("6","9","7","5","8","3","4","0","2","1","10")
layers <- set_names(ord1, ord2)

DEGs_table <- DEGs_table %>%
  select(-layers) %>%
  mutate(clus = str_extract(.$cluster, "\\d+")) %>%
  mutate(cluster = factor(.$cluster, levels = paste0(ord2, "_all"))) %>%
  mutate(Clusters = factor(layers[as.character(.$clus)], levels = ord1)) 

#saveRDS(DATA_sub_s, paste0(result_dir, "DATA_sub_s.RDS"))
#saveRDS(DATA_sub_e, paste0(result_dir, "DATA_sub_e.RDS"))
```

```{r DEGs-SubMuc-vs-epi, include=TRUE, eval=params$re.run}
##############################
# DGEs EPI AGAINST SUBMUCOSA #
##############################
DEGs_table_morf <- DATA_sub$full_subset %>%
  unnest(subset_morf) %>%
  DEGs_fun(., "sp_annot") %>%
  filter(cluster == "epi_all") %>%
  mutate(Regulation = ifelse(avg_log2FC > 0 & p_val_adj < 0.05, "UP",
                            ifelse(avg_log2FC < 0 & p_val_adj < 0.05,"DOWN", "NOT SIG.")))

top_morf <- DEGs_table_morf %>%
  filter(p_val_adj < 0.05) %>%
  {. ->> sig_morf} %>%
  group_by(Regulation) %>%
  top_n(15, abs(avg_log2FC)) %>%
  arrange(desc(avg_log2FC))

summary_morf <- sig_morf %>% 
  summarise(n = n(), .by = "Regulation") %>%
  add_case(Regulation="Total", n=sum(.$n))

KRT_morf <- DEGs_table_morf %>%
  filter(grepl("^KRT", .$gene))
```

```{r Avg-expression}
#################################
# ADD AVERAGE EXPRESSION VALUES #
#################################
# https://github.com/satijalab/seurat/issues/5495

#DATA_sub_s <- readRDS(paste0(result_dir, "DATA_sub_s.RDS"))
#DATA_sub_e <- readRDS(paste0(result_dir, "DATA_sub_e.RDS"))
avg.fun <- function(obj, grs, long=TRUE){
  rowMeans_drop0 <- function (obj, name, filt=NULL) {
    # average expression of spots with values > 0
    obj <- obj[,filt]
    dgCMat <- obj@assays[["RNA"]]@data
    RowInd <- dgCMat@i + 1
    nnz_per_row <- tabulate(RowInd)
    nnz_per_row[nnz_per_row == 0] <- 1  ## just avoid doing 0 / 0
    fix_names <- function(x) gsub("\\s+", name, x)
    tibble(gene = dgCMat@Dimnames[[1]], "Avg_exp_ " = rowSums(dgCMat) / nnz_per_row, .name_repair = fix_names)
  }
  gr <- unique(obj@meta.data[[grs]])
  
  # long format
  Avg_l <- gr %>% set_names() %>%  as.list() %>% 
    map(., ~rowMeans_drop0(obj, .x, obj[[grs]] == .x )) %>%
    {. ->> avg} %>%
    map(., ~rename_with(.x, ~str_replace(.x, "exp_.+", "exp")) ) %>%
    bind_rows(., .id = "groups") %>%
    mutate(groups = as.numeric(.$groups))
  
  # wide format
  Avg_w <- avg %>% purrr::reduce(left_join, by = "gene")
  
  if(long){Avg <- Avg_l}else{Avg <- Avg_w}
  return(Avg)
}
avg_epi <- avg.fun(DATA_sub_e, "Clusters")
avg_epi <- avg.fun(DATA_sub_e, "Clusters", long=F)
avg_SubMuc <- avg.fun(DATA_sub_s, "Clusters", long=F)
avg <- bind_cols(avg_epi, avg_SubMuc)

Avg_all <- function(db, gr, avg=TRUE){
  library(rlang)
   if(avg)
     {db %>% 
      mutate(Avg_all = case_when(!!!parse_exprs(
        paste0('clus == ', gr, '~ rowSums(.[grep("exp_[^', gr, ']", names(.))])/',length(gr) ))) )}
  
  else{db %>% 
      mutate(Avg_diff = case_when(!!!parse_exprs(
        paste0('clus == ', gr, '~ Avg_exp_', gr, '-Avg_all' ))) )}
  }
  

# DEGs_table <- left_join(DEGs_table, avg, by=c("clus"="groups", "gene"))

DEGs_table_epi <- DEGs_table %>%
  filter(grepl(epi_clus, .$clus)) %>%
  left_join(., avg_epi, by=c("gene")) %>%
  Avg_all(., unique(.$clus)) %>%
  Avg_all(., unique(.$clus), avg=FALSE) 

DEGs_table_sub <- DEGs_table %>%
  filter(!(grepl(epi_clus, .$clus))) %>%
  left_join(., avg_SubMuc, by=c("gene")) %>%
  select(-Avg_exp_10) %>%
  Avg_all(., unique(.$clus)) %>%
  Avg_all(., c("8", "4", "0", "3", "1", "2"), avg=FALSE) 
  
```

### Save files
```{r save-DGEs, include=TRUE, eval=params$re.run}
#################
# SAVE CLUSTERS #
#################
write_csv(DEGs_table, paste0(result_dir, "DGEs_clusters_wilcox.csv"))

DEGs_list <- DEGs_table %>% 
  filter(p_val < 0.9) %>%
  split(., ~layers)

write.xlsx(DEGs_list[ord1], keepNA=TRUE, na.string="NA", overwrite=TRUE, 
           file=paste0(result_dir,"Supplemental Table 3",".xlsx"))

###################
# SAVE MORPHOLOGY #
###################
write_csv(DEGs_table_morf, paste0(result_dir, "DGEs_clusters_morf.csv"))

morf_list <- list("DEGs_Epi_vs_SubMuc"=DEGs_table_morf, "Top_15_Up&Down"=top_morf, 
                  "Summary"=summary_morf, "Keratins"=KRT_morf ) 
write.xlsx(morf_list, keepNA=TRUE, na.string="NA", overwrite=TRUE,
           file=paste0(result_dir,"Supplemental Table 2",".xlsx"))
```

### Load allready saved data
```{r get_DEGs, include=FALSE, eval=isFALSE(params$re.run)}
DEGs_table <- read_csv(paste0(result_dir,"DGEs_clusters_wilcox.csv"))
DEGs_table_morf <- read_csv(paste0(result_dir,"DGEs_clusters_morf.csv"))
#DEGs_table_morf <- readxl::read_xlsx(paste0(result_dir,"DGEs_epi_vs_Submuc_wilcox.xlsx"))
# DATA_sub <- readRDS(paste0("../results/06_DGE_condition_st_data/", "DATA_sub.RDS"))
# DATA_sub <- DATA_sub$DATA_sub 
```

```{r add-regulation}
###################
# ADD ANNOTATION #
##################
DEGs_l <- list("Clusters"=DEGs_table, "Morf"=DEGs_table_morf)

DEGs_l <- map(DEGs_l, ~ .x %>%
  mutate(Regulation = ifelse(avg_log2FC > 0 & p_val_adj < 0.05, "UP",
                            ifelse(avg_log2FC < 0 & p_val_adj < 0.05,"DOWN", "NOT SIG."))) %>%
  mutate(Regulation = factor(.$Regulation, levels = c("UP", "DOWN", "NOT SIG.")))
) 

DEGs_table <- DEGs_l[[1]]
DEGs_table_E2lvl <- DEGs_l[[2]]
```

### Volcano plot of DEGs for each cluster
```{r 05a_DEGs_volcano_plot_epi, fig.height=5}
###########################
# VOLCANO PLOT EPITHELIUM #
###########################
DEGs_filt <- DEGs_table %>% 
  filter(grepl(epi_clus, .$clus)) %>%
  filter(p_val < 0.099) 

Volcano.fun_logFC(DEGs_filt, "layers", y.axis="p-value", 
                  up=c(.2, 0.05), down = c(-.2, 0.05)) # labeling: (logFC, p-value)
```

```{r 05b_DEGs_volcano_plot_subMuc, fig.height=6}
###########################
# VOLCANO PLOT SUBMUCOSA #
###########################
DEGs_filt <- DEGs_table %>% 
  filter(!(grepl(epi_clus, .$clus))) %>%
  filter(p_val < 0.099) 

Volcano.fun_logFC(DEGs_filt, "cluster", y.axis="p-value", 
                  up=c(.2, 0.05), down = c(-.2, 0.05)) # labeling: (logFC, p-value)
```

```{r save-morf-DEGs}
#############################
# TOP 15 DOWN/UP DEGS TABLE #
#############################
top30 <- DEGs_table_morf %>%
  mutate(Direction = ifelse(avg_log2FC > 0, "UP", "DOWN")) %>%
  #filter(Direction == "UP") %>%
  #filter(pct.1 > 0.2) %>%
  group_by(Direction) %>%
  
  
  top_n(70, abs(avg_log2FC)) %>%
  top_n(-15, p_val_adj) %>% 
  arrange(avg_log2FC, Direction)


write.xlsx(list("DEGs"=DEGs_filt, "top 15 DEGs by log2FC"=top30), keepNA=TRUE, na.string="NA", overwrite=TRUE,
           file=paste0(result_dir,"DGEs_epi_vs_Submuc_25_25_wilcox",".xlsx"))

```

```{r 05c_dotplott_top40_DEGs_sm_vs_epi, fig.height=40*.3+2, fig.width=4}
#| fig-format: pdf
############################
# MARKER GENES FOR DOTPLOT #
############################
marker_genes_morf <- DEGs_table_morf %>%
  mutate(Direction = ifelse(avg_log2FC > 0, "UP", "DOWN")) %>%
  group_by(Direction) %>%
  top_n(-70, p_val_adj) %>%
  top_n(40, abs(log.pct.diff)) %>%
  {.->> top_40} %>%
  top_n(10, abs(avg_log2FC)) %>%
  arrange(pct.diff, Direction)

# get order of marker genes to seperate clusters
ord <- niceRplots::getcluster(DATA, unique(top_40$gene), "sp_annot")

###################################
# DOTPLOT TOP 10 DOWN AND UP DEGs #
###################################
# dev.new(width=4, height=40*.3+2, noRStudioGD = TRUE) # vertical
#pdf( paste0("./top_20_DEG_subset_epi.pdf"),width=8.27,height=length(ord)/6+2)
par(mfrow=c(1,1), mar=c(7,6,1,5))
plot_dots(DATA, names(sort(ord)), clustering = "sp_annot", 
          show_grid = T,
          main = "top cluster markers",
          cex.main=1,font.main=1,cex.col = 1,srt = 90,cex.row = 1.1)
abline(v=cumsum(c(table(sub( "_.*","",names(table(DATA$sp_annot))))))+0.5)
#dev.off()
```

```{r 05d_dotplott_top5_DEGs_sm_vs_epi, fig.height=2, fig.width=6}
################################
# DOTPLOT MARKER GENE FUNCTION #
################################
library(ggnewscale)
get_order <- function(df, markers, markers_id, value){
  m <- set_names(markers, markers_id)
  value <- enquo(value) 
  
  df <- df %>%
    mutate(marker_id =  map_chr(.$marker, ~names(m)[which(m == as.character(.x))][1])) %>%
    nest(., .by=c(groups)) %>%
    mutate(., data = pmap(., ~filter(..2, marker_id == .[[1]]))) %>%
    mutate(., data = pmap(., ~arrange(..2, desc(!!(value))))) %>%
    unnest(cols = c("data")) %>%
    #mutate(marker = fct_reorder2(marker, Avg, Pct))
    mutate(row_id = cur_group_rows(), .by=c(groups))
  
  return(df[["marker"]]) # use rev() to reorder the dots depending on horizontal/vertical 
}
get_df_for_dotplot.fun <- function(obj, markers, groups="Clusters", gr_col, rect=TRUE, alpha=.2, gr_lvl=NULL){
  col <- c("#EFEDF5", "#DADAEB", "#BCBDDC", "#9E9AC8", "#807DBA", "#6A51A3", "#54278F", "#3F007D") # 
  markers_lvl <- set_names(seq_along(markers), markers)
  gr <- unique(obj[[groups]])[[1]]
  gr_lvl <- set_names(seq_along(gr), if(is.null(gr_lvl)){gr}else{gr_lvl})
  
  df <- obj %>%
    mutate(., FetchData(., vars = c(markers)) ) %>%
    as_tibble() %>%
    select(., .cell, any_of(c(groups, markers))) %>%
    pivot_longer(., cols = -c(".cell", groups), 
                 names_to = "marker", values_to = "values") %>%
    #mutate(Clusters = factor(.$Clusters, levels=names(clus_lvl))) %>%
    group_by(!!sym(groups), marker) %>%
    summarise(Avg = mean(values),
              Avg_ = mean(values)/max(values), # gives the median as a percentage of max value among spots if .5 you have a normal dist.
              Pct = sum(values > 0) / length(values) * 100, .groups="drop") %>%
    #group_by(marker) %>%
    #nest()
    mutate(Avg_p = Avg/max(Avg), .by = "marker") %>%
    mutate(marker = factor(.$marker, levels=get_order(., markers, markers_id, Avg))) %>% 
    mutate(., ymin = gr_lvl[as.character(.[[groups]])]-.5,
              ymax = gr_lvl[as.character(.[[groups]])]+.5) %>%
    mutate(., xmin = markers_lvl[as.character(.$marker)]-0.5,
              xmax = markers_lvl[as.character(.$marker)]+0.5) %>%
    mutate(!!sym(groups) := factor(unlist(.[,1]), levels=names(gr_lvl)))
  
  df_rect <- df %>% select(1, ymin,  ymax) %>% slice_max(., n=1, order_by=ymin, by=groups, with_ties=F)
  
  p <- ggplot(df, aes(x=marker, y=.data[[groups]])) +
    {if(rect == TRUE)
      list(new_scale_fill(),
      
      annotate(geom = "rect", ymin=df_rect$ymin, ymax=df_rect$ymax,
                  xmin=rep(0, length(gr)), xmax=rep(length(markers_lvl)+1, length(gr)),
                  fill = gr_col[1:length(gr)], colour = NA, alpha = alpha),
      coord_cartesian(expand = F) )} +
    
    new_scale_fill() +
    geom_point(aes(size = Pct, fill = Avg_), col = "gray", shape = 21) + #"#6A2595"
    scale_fill_gradientn(colours = col,
                       guide = guide_colorbar(ticks.colour = "black",
                                              frame.colour = "black"),
                       name = "Average\nexpression") +
    
    
    theme_bw() + labs(size="% expressed") +
    guides(size = guide_legend(override.aes = list(color = "black") )) +
    theme(axis.text.x = element_text(size=12, angle=45, hjust=1, color="black"),
          axis.text.y = element_text(size=12, color="black",hjust=.9 ),
          legend.text = element_text(size=10),
          legend.margin=margin(0,0,0,0), #t,r,b,l
          legend.box.margin=margin(0,-6,-10,-8),
          axis.title = element_blank(),
          #panel.grid.major = element_blank(), 
          #panel.grid.minor = element_blank(), 
          axis.ticks.y = element_blank(),
          plot.margin = unit(c(.2,.2,.2,.2),units = "cm") #trbl
          ) 
  return(p)
}

###################################
# DOTPLOT TOP 10 DOWN AND UP DEGs #
###################################
# dev.new(width=6, height=2, noRStudioGD = TRUE)
gr_col <- c("#F0B6AD","#BACCE5")
markers <-  marker_genes_morf$gene
markers_id <- rep("epi", length(markers))
groups="sp_annot"
(plot <- get_df_for_dotplot.fun(DATA, markers, groups="sp_annot", gr_col, rect=T ) + theme(legend.position="top")) 

ggsave(paste0("./Figures/05/","Marker_genes_morf", ".pdf"), plot,  width = 6, height = 2)
```

### Identify marker genes to seperate epithelial clusters
```{r Get-best-markrs_epi}
#######################################
# FILTER BY P-VAL logFC AND pct.diff #
######################################
filter_top.fun <- function(nested_df, n, var, min_pct = 0){
  if(is.na(n[4])){n[4] = min_pct}
  res <- nested_df %>% # res <- top20_epi$data[[3]] %>%
    #mutate(p_val_adj = 1.000-.$p_val_adj) %>%
    filter(p_val_adj < 0.05) %>%
    arrange(desc(avg_log2FC)) %>%
    filter(pct.1 > n[4]) %>%
    top_n(., n=n[1], abs(!!sym(var[1]))) %>%
    arrange(desc(avg_log2FC)) %>%
    #top_n(., n=n[2], Avg_exp) %>%
    top_n(., n=n[2], abs(!!sym(var[2]))) %>%
    top_n(., n=n[3], abs(!!sym(var[3]))) 
  print(res)
  return(res[1:n[3],])
}

sel_epi <- list(
  #Sup_1 = list(n=c(-40, 30, 5),var=c("p_val_adj", "pct.diff", "Avg_exp", "log.pct.diff")),
  "Superficial" = list(n=c(-40, 50, 5),var=c("p_val_adj", "pct.diff", "log.pct.diff")),
  #Sup_1 = list(n=c(-100, 70, 5, 0.5),var=c("pct.diff", "p_val_adj", "log.pct.diff")),
  "Upper IM" = list(n=c(-40, 50, 5),var=c("p_val_adj", "pct.diff", "log.pct.diff")),
  #Sup_2 = list(n=c(20, 5, 5),var=c("avg_log2FC", "log.pct.diff", "log.pct.diff")),
  "Lower IM" = list(n=c(25, 7, 5, .5),var=c("avg_log2FC", "log.pct.diff", "log.pct.diff")),
  "Basal" = list(n=c(10, 5, 5, .8),var=c("avg_log2FC", "log.pct.diff", "log.pct.diff")) )

# Identify the top genes that have a high difference in expression between the clusters
top20_epi <- DEGs_table %>%
  filter(grepl(epi_clus, .$clus)) %>%
  arrange(match(Clusters, names(sel_epi))) %>%
  mutate(Direction = ifelse(avg_log2FC > 0, "UP", "DOWN")) %>%
  filter(Direction == "UP") %>%
  #filter(!(pct.2 > 0.2)) %>%
  nest(., .by = "Clusters") %>%
  mutate(data = pmap(., ~filter_top.fun(..2, n=sel_epi[[..1]]$n, var=sel_epi[[..1]]$var)) ) %>%
  unnest(., cols=c("data"))
  
```

### Dotplot of epithelial marker genes 
```{r 05e_Marker_genes_epi, fig.height=3, fig.width=8}
###################################
# DOTPLOT TOP 10 MARKER GENES EPI #
###################################
# dev.new(width=6.5, height=3, noRStudioGD = TRUE)
gr_col <-c("#EB5253","#FA9F50","#D7A1FF","#984EA3")
markers <- top20_epi$gene
markers_id <- top20_epi$Clusters
#markers <- c("FLG","DUOXA2","DHRS9","SPRR2B","KPRP","LINC02487","CLCA4","CD177","SERPINB2","RPTN","TGM1","TMEM184A","EPHA1","FAM83G","ALOX15B","GALNT5","MUC22","COL7A1","STMN1","FGFR2","LAMA5","MEGF6","MIR205HG")
#markers_id <- c(rep("Sup_1",5), rep("Sup_2",6),rep("Basal_2",6),rep("Basal_1",6))
groups="layers"
gr_lvl <- rev(c("Superficial", "Upper IM", "Lower IM", "Basal"))
(plot <- DATA %>%
    filter(grepl(epi_clus, .$Clusters) & sp_annot == "epi") %>%
    get_df_for_dotplot.fun(., markers, groups="layers", gr_col, rect=T, gr_lvl=gr_lvl)  +
    theme(legend.position="top") + guides(fill = guide_colourbar(barwidth = 7, barheight = .5)) # horizontal
    #coord_flip(expand=F) + theme(legend.position="right") + guides(fill = guide_colourbar(barwidth = .5, barheight = 10)) # vertical
  )
# dev.new(width=8, height=3, noRStudioGD = TRUE) # horizontal
# dev.new(width=4, height=5.5, noRStudioGD = TRUE) # vertical
# ggsave(paste0("./Figures/05/","Marker_genes_epi_h", ".pdf"), plot,  width = 6, height = 3)
# ggsave(paste0("./Figures/05/","Marker_genes_epi_v", ".pdf"), plot,  width = 4, height = 5.5)
```

### Identify marker genes to seperate Submucosal clusters
```{r Get-best-markrs_SubMuc}
#######################################
# FILTER BY P-VAL logFC AND pct.diff #
######################################
filter_top.fun <- function(nested_df, n, var, pct = 0.01, filt=T){
  if(is.na(n[4])){n[4] = pct}
  res <- nested_df %>% # res <- top20_sub$data[[3]] %>%
    #mutate(p_val_adj = 1.000-.$p_val_adj) %>%
    #filter(Avg_diff > 0) %>%
    #filter(pct.2 < n[4]) %>%
    top_n(., n=n[1], abs(!!sym(var[1]))) %>%
    top_n(., n=n[2], abs(!!sym(var[2]))) %>%
    {if(filt) filter(., between(pct.1, 0.8,0.3) | between(pct.2, 0.1,0.4)) else .} %>%
    #arrange(desc(Avg_exp_3)) %>%
    arrange(log.pct.diff) %>%
    top_n(., n=n[3], abs(!!sym(var[3]))) 
  print(res)
  return(res[1:n[3],])
}
sel_s <- list(
  "8" = list(n=c(-100, 50, 5),var=c("p_val_adj", "pct.diff", "log.pct.diff"), filt=F),
  "3" = list(n=c(-90, 30, 5),var=c("p_val_adj", "avg_log2FC", "avg_log2FC"), filt=F),
  #"3" = list(filt=F, n=c(-150, 100, 5),var=c("p_val_adj", "pct.diff", "log.pct.diff")),
  "4" = list( n=c(-100, 50, 5),var=c("p_val_adj", "pct.diff", "log.pct.diff"), filt=F),
  "0" = list( n=c( -50, 30, 5, 0.2),var=c( "p_val_adj", "avg_log2FC", "Avg_diff", "log.pct.diff"), filt=T),
  #"0" = list( n=c( -50, 30, 5, 0.2),var=c( "p_val_adj", "avg_log2FC", "log.pct.diff"), filt=T),
  "2" = list(n=c(-90, 20, 5, 0.1),var=c("p_val_adj", "avg_log2FC", "avg_log2FC", "pct.diff"), filt=T),
  "1" = list( n=c(-90, 50, 5, 0.1),var=c("p_val_adj", "pct.diff", "log.pct.diff"), filt=T)
  )

# Identify the top genes that have a high difference in expression between the clusters
top20_sub <- DEGs_table_sub %>%
  filter(!(grepl(epi_clus, .$clus))) %>%
  
  filter(!(clus == "10")) %>%
  mutate(Direction = ifelse(avg_log2FC > 0, "UP", "DOWN")) %>%
  filter(Direction == "UP") %>%
  #filter(!(pct.2 > 0.2)) %>%
  nest(., .by = "Clusters") %>%
  arrange(match(Clusters, names(sel_s))) %>%
  mutate(data = pmap(., ~filter_top.fun(..2, n=sel_s[[..1]]$n, var=sel_s[[..1]]$var), filt=sel_s[[..1]]$filt)) %>%
  unnest(., cols=c("data"))
  dput(top20_sub$gene)
```

### Dotplot of submucosal marker genes 
```{r 05f_Marker_genes_SubMuc, fig.height=4, fig.width=9}
######################################
# DOTPLOT TOP 10 MARKER GENES SUBMUC #
######################################
# dev.new(width=6.5, height=3, noRStudioGD = TRUE)
gr_col <- c("#FBAAB1","#FFDAB8","#F1D2FF","#E6B3E9","#8BCFCE","#92DBFF","#ABC9E1","#E5C264","#E1E2A4","#BBD99B","#FDFFD3")[5:10]
markers <- top20_sub$gene
markers_id <- top20_sub$clus
groups="Clusters"
gr_lvl <- c("8","3","4","0","2","1")
library(ggnewscale)
(plot <- DATA %>%
    filter(grepl("^8|^3|^4|^0|^1$|^2", .$Clusters) & sp_annot == "SubMuc") %>%
    mutate(Clusters = factor(.$Clusters, levels = gr_lvl)) %>%
    get_df_for_dotplot.fun(., markers, groups="Clusters", gr_col, rect=T, alpha = .3, gr_lvl=gr_lvl) + 
    #theme(legend.position="top") + guides(fill = guide_colourbar(barwidth = 10, barheight = .5)) # vertical
    coord_flip(expand = F) + theme(legend.position="right") + guides(fill = guide_colourbar(barwidth = .5, barheight = 10)) # horizontal
  )

# dev.new(width=5.2, height=7, noRStudioGD = TRUE) # horizontal
# dev.new(width=7.5, height=3.3, noRStudioGD = TRUE) # vertical
# ggsave(paste0("./Figures/05/","Marker_genes_SubMuc_v", ".pdf"), plot,  width = 7.5, height = 3.3)
# ggsave(paste0("./Figures/05/","Marker_genes_SubMuc_h", ".pdf"), plot,  width=5, height = 6.5)
```


```{r}
######################################
# SAVE CLUSTER DEGs AND MARKER GENES #
######################################
DEGs_list <- DEGs_table %>% 
  arrange(Clusters) %>%
  filter(p_val < 0.9) %>%
  split(., ~Clusters)

markers <- list(top20_epi, top20_sub) %>% map(., ~select(.x, -clus, -starts_with("Avg"))) %>% bind_rows()

append( list("top_markers_all_clusters"=markers), DEGs_list ) %>%
  write.xlsx(., keepNA=TRUE, na.string="NA", overwrite=TRUE, 
           file=paste0(result_dir,"Supplemental Table 3",".xlsx"))
```


```{r 05g_dotplott_top_DEGs_epi, fig.height=40*.25+2, fig.width=4}
#| fig-format: pdf

# Identify the top genes that have a high difference in expression between the clusters
top_epi <- DEGs_table %>%
  filter(grepl(epi_clus, .$clus )) %>%
  filter(avg_log2FC > 0) %>%
  group_by(layers) %>%
  top_n(-60, p_val_adj) %>%
  top_n(20, pct.diff) %>%
  slice_max(., n=10, order_by=log.pct.diff, with_ties = F )

# Select marker genes to seperate clusters
DAT <- DATA  %>% 
  filter(grepl(epi_clus, .$Clusters )) %>%
  mutate(layers = factor(.$layers, levels=c("Sup_1", "Sup_2", "Basal_2", "Basal_1") ) )
  
ord <- niceRplots::getcluster(DAT, unique(top_epi$gene), "layers")

######################################
# DOTPLOT TOP 10 MARKERS PER CLUSTER #
######################################
# dev.new(width=6, height=40*.2+2, noRStudioGD = TRUE)
#pdf( paste0("./Figures/05/","top_20_DEG_subset_epi_pval.pdf"), width=8.27,height=length(ord)/6+2, bg = "transparent")
par(mfrow=c(1,1), mar=c(7,6,1,5))
plot_dots(DAT, names(sort(ord)), clustering = "layers", 
          show_grid = T,
          main = "top cluster markers",
          cex.main=1, font.main=1,
          cex.col = 1, srt = 90,cex.row = 1.1)
abline(v=cumsum(c(table(sub( "_.*","",names(table(DAT$layers))))))+0.5)
#dev.off()
```

```{r 05h_dotplott_top_DEGs_SubMuc, fig.height=60*.2+2, fig.width=6}
#| fig-format: pdf

# Identify the top genes that have a high difference in expression between the clusters
top_sub <- DEGs_table %>%
  filter(!(grepl(epi_clus, .$clus ))) %>%
  filter(avg_log2FC > 0) %>%
  filter(pct.1 > .4) %>%
  group_by(layers) %>%
  top_n(-60, p_val_adj) %>%
  top_n(20, pct.diff) %>%
  slice_max(., n=10, order_by=log.pct.diff, with_ties = F )

# Select marker genes to seperate clusters
DAT <- DATA  %>% 
  filter(!(grepl(epi_clus, .$Clusters ))) %>%
  mutate(layers = factor(.$layers, levels=c("8","3","4","2","1","0","10") ) )
  
ord <- niceRplots::getcluster(DAT, top_sub$gene, "layers")

######################################
# DOTPLOT TOP 10 MARKERS PER CLUSTER #
######################################
# dev.new(width=6, height=40*.25+2, noRStudioGD = TRUE)
#pdf( paste0("./Figures/05/","top_20_DEG_subset_epi_pval.pdf"), width=8.27,height=length(ord)/6+2, bg = "transparent")
par(mfrow=c(1,1), mar=c(7,6,1,5))
plot_dots(DAT, names(sort(ord)), clustering = "layers", 
          show_grid = T,
          main = "top cluster markers",
          cex.main=1, font.main=1,
          cex.col = 1, srt = 90,cex.row = 1.1)
abline(v=cumsum(c(table(sub( "_.*","",names(table(DAT$layers))))))+0.5)
#dev.off()
```

```{r 05i_violin_DEGs_plot, fig.height=ceiling(length(top20_epi$gene)/3)*2, fig.width=7}
# dev.new(width=6.6929133858, height=round(length(unique(top20_morf$epi$gene))/3)*2, noRStudioGD = TRUE) 
################################
# VIOLIN PLOT BEFORE FILTERING #
################################
col <- c("#E41A1C","#FF7F00","#C77CFF","#984EA3")
feature <-  top20_epi$gene

DAT <- DATA  %>%
  filter(grepl(epi_clus, .$Clusters )) %>%
  mutate(., FetchData(., vars = c(feature)) ) 
p <-  map(feature, ~violin.fun(DAT, facet="layers", .x, fill="layers", col_pal=col))

plot_grid(plotlist=p, nrow = ceiling(length(feature)/3), ncol = 3, byrow = T)
```

```{r 05j_violin_DEGs_sm_vs_epi, fig.height=ceiling(length(marker_genes_morf$gene)/3)*2, fig.width=10}
# dev.new(width=6.6929133858, height=round(length(unique(top20_sm_vs_epi$gene))/2)*2, noRStudioGD = TRUE) 
# dev.new(width=6.6929133858, height=round(length(feature)/2)*2, noRStudioGD = TRUE) 
################################
# VIOLIN PLOT BEFORE FILTERING #
################################
col <- c("#E41A1C","#FF7F00","#C77CFF","#984EA3", "#CD9600","#00BFC4","#F8766D","#7CAE00", "#00A9FF", "#377EB8","#FFFF33")
feature <-  marker_genes_morf$gene

DAT <- DATA  %>%
  mutate(., FetchData(., vars = c(feature)) ) 
p <-  map(feature, ~violin.fun(DAT, facet="layers", .x, fill="layers", col_pal=col))

plot_grid(plotlist=p, nrow = ceiling(length(feature)/2), ncol = 2, byrow = T)
```

```{r 05k_barplot_top25_DGEs, fig.width=10, fig.height=6}
###################################
# BARPLOT TOP 10 DOWN AND UP DEGs #
###################################
# dev.new(width=10, height=6, noRStudioGD = TRUE) 
top20 <- bind_rows(top_epi, top_sub)

par(mfrow=c(2, 5), mar = c(4, 6, 3, 1))
for (i in unique(top20$layers)[1:10]) {
    barplot(sort(setNames(top20$avg_log2FC, top20$gene)[top20$layers == i], F),
        horiz = T, las = 1, main = paste0(i, " vs. rest"), border = "white", yaxs = "i")
    abline(v = c(0, 0.25), lty = c(1, 2))
}
```

```{r 05l_UMAP_top10_markergenes_plot, fig.width=8.5, fig.height=6.5}
#| fig-format: png
# dev.new(width=8.5, height=6.5, noRStudioGD = TRUE) 
################################
# UMAP TOP 10 DOWN AND UP DEGs #
################################
col <- c("#EFEDF5", "#DADAEB", "#BCBDDC", "#9E9AC8", "#807DBA", "#6A51A3", "#54278F", "#3F007D") # Purples
clus <-   c("#CD9600","#7CAE00","#e0e067","#00A9FF","#377EB8","#984EA3","#E41A1C","#C77CFF","#00BFC4","#FF7F00","#FFFF33")

clus_plot <- plot_clusters.fun(DATA, red = "umapharmony", color = clus,
                               dot_size = 0.2, cluster="Clusters") + theme_void() + NoLegend()

grid_genes <- function(plot_list, title){
  title <- ggdraw() + draw_label(paste0("Top 10 Markers for Cluster ", title), fontface='bold')
  g <- plot_grid(plotlist = plot_list,
            ncol = 4)
  g_ <- plot_grid(title, g, ncol=1, rel_heights=c(0.1, 1))
  return(g_)
}

top20_gr <- top20 %>%
  ungroup() %>%
  split(., f = .$layers)

plots <- imap(top20_gr, "gene") %>%
  map(., map, ~plot_genes.fun(DATA, red = "umapharmony", .x, col = col, lable = "Clusters", point_size = .2))

cluster_markers <- plots %>%
  map(., ~c("Clusters"=list(clus_plot), .x)) %>%
  imap(., ~grid_genes(.x, .y ))

cluster_markers[[10]]

# imap(cluster_markers, ~ggsave(paste0(marker_dir,"Marker_genes_cluster_", .y, ".jpg"), plot=.x, height = 6.5, width = 8.5))
pdf(paste0("./Figures/05/","Marker_genes_UMAP", ".pdf"), height = 6.5, width = 8.5)
cluster_markers
dev.off()
```

## Session info
```{r}
sessionInfo()
```
